+++ lib/handlers.dart
@@ -4,6 +4,6 @@ export './src/handlers/blank.dart';
 
 export './src/handlers/direct.dart' show direct;
 export './src/handlers/parser.dart' show HandlerType;
-export './src/handlers/json.dart' show JsonCTX, json;
-export './src/handlers/form.dart' show FormCTX, form;
-export './src/handlers/iform.dart' show iFormCTX, iForm;
+export './src/handlers/json.dart' show JsonCTX;
+export './src/handlers/form.dart' show FormCTX;
+export './src/handlers/iform.dart' show iFormCTX;